Refactor: extract mixins from Fish, deduplicate poker algorithms
Architecture cleanup reducing code duplication and improving modularity:

- Extract EnergyManagementMixin, MortalityMixin, ReproductionMixin from
  Fish entity (1273 -> 797 lines) into core/entities/mixins/
- Deduplicate poker algorithm patterns: extract _flee_from_predator,
  _seek_food_or_idle, _seek_fish_or_none shared helpers (886 -> 763 lines)
- Fix duplicate section comment and stale comment in ecosystem.py
- Remove unused import (ReproductionEvent) and redundant local import
  (MemoryType) from fish.py
- Update architecture guard tests to allow Fish mixin files

All 1569 tests pass.
