+++ core/algorithms/poker.py
@@ -16,6 +16,7 @@ Performance optimizations:
 - Caches fish position coordinates
 """
 
+import math
 import random
 from dataclasses import dataclass
 from typing import TYPE_CHECKING, Any, Dict, List, Optional, Tuple
@@ -25,6 +26,11 @@ from core.algorithms.base import BehaviorAlgorithm
 if TYPE_CHECKING:
     from core.entities import Fish
     from core.world import World
+
+
+# =========================================================================
+# Shared spatial query helpers
+# =========================================================================
 
 
 def _find_nearest_fish_spatial(fish: "Fish", radius: float) -> Tuple[Optional["Fish"], float]:
@@ -97,6 +103,93 @@ def _get_nearby_fish_spatial(fish: "Fish", radius: float) -> List["Fish"]:
     return [f for f in nearby if isinstance(f, FishClass) and f.fish_id != fish_id]
 
 
+# =========================================================================
+# Shared behavior helpers (eliminate duplication across poker algorithms)
+# =========================================================================
+
+
+def _flee_from_predator(
+    algo: BehaviorAlgorithm,
+    fish: "Fish",
+    fish_x: float,
+    fish_y: float,
+    danger_dist_sq: float = 14400,
+    flee_speed: float = 1.2,
+) -> Optional[Tuple[float, float]]:
+    """Check for nearby predator and return flee velocity if threatened.
+
+    This consolidates the predator-evasion pattern used by all poker algorithms.
+    Uses squared distances to avoid sqrt overhead when predator is far away.
+
+    Args:
+        algo: The algorithm instance (for _find_nearest helper)
+        fish: The fish checking for threats
+        fish_x: Cached fish x position
+        fish_y: Cached fish y position
+        danger_dist_sq: Squared distance threshold for fleeing (default 120^2)
+        flee_speed: Speed multiplier when fleeing (default 1.2)
+
+    Returns:
+        (vx, vy) flee velocity if threatened, or None if safe
+    """
+    from core.entities import Crab
+
+    nearest_predator = algo._find_nearest(fish, Crab)
+    if nearest_predator:
+        dx = nearest_predator.pos.x - fish_x
+        dy = nearest_predator.pos.y - fish_y
+        dist_sq = dx * dx + dy * dy
+        if dist_sq < danger_dist_sq:
+            dist = math.sqrt(dist_sq)
+            if dist > 0:
+                return -dx / dist * flee_speed, -dy / dist * flee_speed
+            return 0.0, 0.0
+    return None
+
+
+def _seek_food_or_idle(
+    algo: BehaviorAlgorithm, fish: "Fish", speed: float = 1.0
+) -> Tuple[float, float]:
+    """Seek nearest food at given speed, or return (0, 0) if no food found.
+
+    This consolidates the food-seeking fallback pattern used by all poker algorithms.
+
+    Args:
+        algo: The algorithm instance (for _find_nearest_food helper)
+        fish: The fish seeking food
+        speed: Speed multiplier for movement toward food
+
+    Returns:
+        (vx, vy) direction toward food, or (0, 0) if no food found
+    """
+    nearest_food = algo._find_nearest_food(fish)
+    if nearest_food:
+        direction = algo._safe_normalize(nearest_food.pos - fish.pos)
+        return direction.x * speed, direction.y * speed
+    return 0.0, 0.0
+
+
+def _seek_fish_or_none(
+    algo: BehaviorAlgorithm, fish: "Fish", radius: float, speed: float = 1.0
+) -> Optional[Tuple[float, float]]:
+    """Seek nearest other fish within radius. Returns direction or None.
+
+    Args:
+        algo: The algorithm instance (for _safe_normalize helper)
+        fish: The fish seeking others
+        radius: Search radius for spatial query
+        speed: Speed multiplier for movement
+
+    Returns:
+        (vx, vy) direction toward nearest fish, or None if no fish nearby
+    """
+    nearest_fish, _ = _find_nearest_fish_spatial(fish, radius)
+    if nearest_fish is not None:
+        direction = algo._safe_normalize(nearest_fish.pos - fish.pos)
+        return direction.x * speed, direction.y * speed
+    return None
+
+
 @dataclass
 class PokerChallenger(BehaviorAlgorithm):
     """Actively seeks out other fish for poker games."""
@@ -121,43 +214,20 @@ class PokerChallenger(BehaviorAlgorithm):
         return cls(rng=rng)
 
     def execute(self, fish: "Fish") -> Tuple[float, float]:
-        import math
-
-        from core.entities import Crab
-
         fish_x, fish_y = fish.pos.x, fish.pos.y
 
-        # First check for predators - survival comes first
-        nearest_predator = self._find_nearest(fish, Crab)
-        if nearest_predator:
-            dx = nearest_predator.pos.x - fish_x
-            dy = nearest_predator.pos.y - fish_y
-            dist_sq = dx * dx + dy * dy
-            if dist_sq < 14400:  # 120^2
-                dist = math.sqrt(dist_sq)
-                if dist > 0:
-                    # Flee from predator (normalize and flip direction)
-                    return -dx / dist * 1.2, -dy / dist * 1.2
-                return 0, 0
+        flee = _flee_from_predator(self, fish, fish_x, fish_y)
+        if flee is not None:
+            return flee
 
         # Only seek poker if we have enough energy
         if fish.energy < self.parameters["min_energy_to_challenge"]:
-            # Low energy - seek food instead
-            nearest_food = self._find_nearest_food(fish)
-            if nearest_food:
-                direction = self._safe_normalize(nearest_food.pos - fish.pos)
-                return direction.x * 0.8, direction.y * 0.8
-            return 0, 0
+            return _seek_food_or_idle(self, fish, speed=0.8)
 
-        # OPTIMIZATION: Use spatial query instead of get_agents_of_type
-        challenge_radius = self.parameters["challenge_radius"]
-        nearest_fish, dist_sq = _find_nearest_fish_spatial(fish, challenge_radius)
-
-        if nearest_fish is not None:
-            # Move toward the fish for poker
-            direction = self._safe_normalize(nearest_fish.pos - fish.pos)
-            speed = self.parameters["challenge_speed"]
-            return direction.x * speed, direction.y * speed
+        # Seek nearest fish for poker challenge
+        result = _seek_fish_or_none(self, fish, self.parameters["challenge_radius"], self.parameters["challenge_speed"])
+        if result is not None:
+            return result
 
         # No fish nearby - wander
         return self.rng.uniform(-0.5, 0.5), self.rng.uniform(-0.5, 0.5)
@@ -187,57 +257,35 @@ class PokerDodger(BehaviorAlgorithm):
         return cls(rng=rng)
 
     def execute(self, fish: "Fish") -> Tuple[float, float]:
-        from core.entities import Crab
+        fish_x, fish_y = fish.pos.x, fish.pos.y
 
-        # First check for predators
-        nearest_predator = self._find_nearest(fish, Crab)
-        if nearest_predator:
-            dx = nearest_predator.pos.x - fish.pos.x
-            dy = nearest_predator.pos.y - fish.pos.y
-            dist_sq = dx * dx + dy * dy
-            if dist_sq < 14400:  # 120^2
-                import math
+        flee = _flee_from_predator(self, fish, fish_x, fish_y)
+        if flee is not None:
+            return flee
 
-                dist = math.sqrt(dist_sq)
-                if dist > 0:
-                    return -dx / dist * 1.2, -dy / dist * 1.2
-                return 0, 0
-
-        # OPTIMIZATION: Use spatial query with avoidance radius
+        # Calculate avoidance vector from nearby fish
         avoidance_radius = self.parameters["avoidance_radius"]
         avoidance_radius_sq = avoidance_radius * avoidance_radius
         other_fish = _get_nearby_fish_spatial(fish, avoidance_radius)
 
-        # Calculate avoidance vector using raw floats (avoid Vector2 allocations)
         avoidance_x = 0.0
         avoidance_y = 0.0
         fish_nearby = 0
-        fish_x, fish_y = fish.pos.x, fish.pos.y
 
         for other in other_fish:
             dx = other.pos.x - fish_x
             dy = other.pos.y - fish_y
             dist_sq = dx * dx + dy * dy
 
-            if dist_sq > 0 and dist_sq < avoidance_radius_sq:
-                import math
-
+            if 0 < dist_sq < avoidance_radius_sq:
                 distance = math.sqrt(dist_sq)
-                # Avoid this fish - direction away from other
                 inv_dist = 1.0 / distance
-                avoid_x = -dx * inv_dist
-                avoid_y = -dy * inv_dist
-                # Stronger avoidance for closer fish
                 strength = (avoidance_radius - distance) / avoidance_radius
-                avoidance_x += avoid_x * strength
-                avoidance_y += avoid_y * strength
+                avoidance_x += -dx * inv_dist * strength
+                avoidance_y += -dy * inv_dist * strength
                 fish_nearby += 1
 
-        # If fish are nearby, move away from them
         if fish_nearby > 0:
-            # Normalize avoidance vector
-            import math
-
             avoid_len = math.sqrt(avoidance_x * avoidance_x + avoidance_y * avoidance_y)
             if avoid_len > 0:
                 avoidance_x /= avoid_len
@@ -245,17 +293,16 @@ class PokerDodger(BehaviorAlgorithm):
 
             speed = self.parameters["avoidance_speed"]
 
-            # Still seek food while avoiding
+            # Blend avoidance with food seeking
             nearest_food = self._find_nearest_food(fish)
             if nearest_food:
                 food_dx = nearest_food.pos.x - fish_x
                 food_dy = nearest_food.pos.y - fish_y
                 food_dist_sq = food_dx * food_dx + food_dy * food_dy
-                if food_dist_sq < 40000 and food_dist_sq > 0:  # 200^2
+                if 0 < food_dist_sq < 40000:  # 200^2
                     food_dist = math.sqrt(food_dist_sq)
                     food_x = food_dx / food_dist
                     food_y = food_dy / food_dist
-                    # Blend avoidance with food seeking
                     final_x = avoidance_x * 0.7 + food_x * 0.3
                     final_y = avoidance_y * 0.7 + food_y * 0.3
                     final_len = math.sqrt(final_x * final_x + final_y * final_y)
@@ -265,15 +312,7 @@ class PokerDodger(BehaviorAlgorithm):
             return avoidance_x * speed, avoidance_y * speed
 
         # No fish nearby - focus on food
-        nearest_food = self._find_nearest_food(fish)
-        if nearest_food:
-            direction = self._safe_normalize(nearest_food.pos - fish.pos)
-            return (
-                direction.x * self.parameters["food_priority"],
-                direction.y * self.parameters["food_priority"],
-            )
-
-        return 0, 0
+        return _seek_food_or_idle(self, fish, speed=self.parameters["food_priority"])
 
 
 @dataclass
@@ -300,52 +339,29 @@ class PokerGambler(BehaviorAlgorithm):
         return cls(rng=rng)
 
     def execute(self, fish: "Fish") -> Tuple[float, float]:
-        import math
-
-        from core.entities import Crab
-
         fish_x, fish_y = fish.pos.x, fish.pos.y
 
-        # Check for predators
-        nearest_predator = self._find_nearest(fish, Crab)
-        if nearest_predator:
-            dx = nearest_predator.pos.x - fish_x
-            dy = nearest_predator.pos.y - fish_y
-            dist_sq = dx * dx + dy * dy
-            if dist_sq < 14400:  # 120^2
-                dist = math.sqrt(dist_sq)
-                if dist > 0:
-                    return -dx / dist * 1.2, -dy / dist * 1.2
-                return 0, 0
+        flee = _flee_from_predator(self, fish, fish_x, fish_y)
+        if flee is not None:
+            return flee
 
         energy_ratio = fish.energy / fish.max_energy
 
         # If energy is high, gamble!
         if energy_ratio > self.parameters["high_energy_threshold"]:
-            # OPTIMIZATION: Use spatial query
-            nearest_fish, _ = _find_nearest_fish_spatial(fish, 200)
-            if nearest_fish is not None:
-                direction = self._safe_normalize(nearest_fish.pos - fish.pos)
-                speed = self.parameters["challenge_speed"]
-                return direction.x * speed, direction.y * speed
+            result = _seek_fish_or_none(self, fish, 200, self.parameters["challenge_speed"])
+            if result is not None:
+                return result
 
         # If energy is medium, balance poker with food
         elif energy_ratio > 0.3:
-            # 50/50 chance between seeking fish or food
             if self.rng.random() < 0.5:
-                # OPTIMIZATION: Use spatial query
-                nearest_fish, _ = _find_nearest_fish_spatial(fish, 200)
-                if nearest_fish is not None:
-                    direction = self._safe_normalize(nearest_fish.pos - fish.pos)
-                    return direction.x * 0.8, direction.y * 0.8
+                result = _seek_fish_or_none(self, fish, 200, 0.8)
+                if result is not None:
+                    return result
 
         # Low energy - focus on food
-        nearest_food = self._find_nearest_food(fish)
-        if nearest_food:
-            direction = self._safe_normalize(nearest_food.pos - fish.pos)
-            return direction.x, direction.y
-
-        return 0, 0
+        return _seek_food_or_idle(self, fish)
 
 
 @dataclass
@@ -373,45 +389,22 @@ class SelectivePoker(BehaviorAlgorithm):
         return cls(rng=rng)
 
     def execute(self, fish: "Fish") -> Tuple[float, float]:
-        import math
-
-        from core.entities import Crab
-
         fish_x, fish_y = fish.pos.x, fish.pos.y
 
-        # Check for predators
-        nearest_predator = self._find_nearest(fish, Crab)
-        if nearest_predator:
-            dx = nearest_predator.pos.x - fish_x
-            dy = nearest_predator.pos.y - fish_y
-            dist_sq = dx * dx + dy * dy
-            if dist_sq < 14400:  # 120^2
-                dist = math.sqrt(dist_sq)
-                if dist > 0:
-                    return -dx / dist * 1.2, -dy / dist * 1.2
-                return 0, 0
+        flee = _flee_from_predator(self, fish, fish_x, fish_y)
+        if flee is not None:
+            return flee
 
         energy_ratio = fish.energy / fish.max_energy
 
         # Only seek poker in the "sweet spot" energy range
         if self.parameters["min_energy_ratio"] < energy_ratio < self.parameters["max_energy_ratio"]:
-            # Be selective - only challenge sometimes
             if self.rng.random() < self.parameters["selectivity"]:
-                # OPTIMIZATION: Use spatial query with max distance 150
-                nearest_fish, dist_sq = _find_nearest_fish_spatial(fish, 150)
+                result = _seek_fish_or_none(self, fish, 150, self.parameters["challenge_speed"])
+                if result is not None:
+                    return result
 
-                if nearest_fish is not None:
-                    direction = self._safe_normalize(nearest_fish.pos - fish.pos)
-                    speed = self.parameters["challenge_speed"]
-                    return direction.x * speed, direction.y * speed
-
-        # Default to food seeking
-        nearest_food = self._find_nearest_food(fish)
-        if nearest_food:
-            direction = self._safe_normalize(nearest_food.pos - fish.pos)
-            return direction.x * 0.9, direction.y * 0.9
-
-        return 0, 0
+        return _seek_food_or_idle(self, fish, speed=0.9)
 
 
 @dataclass
@@ -437,36 +430,20 @@ class PokerOpportunist(BehaviorAlgorithm):
         return cls(rng=rng)
 
     def execute(self, fish: "Fish") -> Tuple[float, float]:
-        import math
-
-        from core.entities import Crab
-
         fish_x, fish_y = fish.pos.x, fish.pos.y
 
-        # Check for predators
-        nearest_predator = self._find_nearest(fish, Crab)
-        if nearest_predator:
-            dx = nearest_predator.pos.x - fish_x
-            dy = nearest_predator.pos.y - fish_y
-            dist_sq = dx * dx + dy * dy
-            if dist_sq < 14400:  # 120^2
-                dist = math.sqrt(dist_sq)
-                if dist > 0:
-                    return -dx / dist * 1.2, -dy / dist * 1.2
-                return 0, 0
+        flee = _flee_from_predator(self, fish, fish_x, fish_y)
+        if flee is not None:
+            return flee
 
         # Look for both food and fish opportunities
         nearest_food = self._find_nearest_food(fish)
+        nearest_fish, _ = _find_nearest_fish_spatial(fish, self.parameters["opportunity_radius"])
 
-        # OPTIMIZATION: Use spatial query
-        opportunity_radius = self.parameters["opportunity_radius"]
-        nearest_fish, _ = _find_nearest_fish_spatial(fish, opportunity_radius)
-
-        # Use raw floats for vectors to avoid allocations
+        # Calculate normalized direction vectors
         food_x, food_y = 0.0, 0.0
         poker_x, poker_y = 0.0, 0.0
 
-        # Calculate food attraction
         if nearest_food:
             dx = nearest_food.pos.x - fish_x
             dy = nearest_food.pos.y - fish_y
@@ -476,7 +453,6 @@ class PokerOpportunist(BehaviorAlgorithm):
                 food_x = dx / dist
                 food_y = dy / dist
 
-        # Calculate poker attraction (nearest fish)
         if nearest_fish is not None:
             dx = nearest_fish.pos.x - fish_x
             dy = nearest_fish.pos.y - fish_y
@@ -486,26 +462,22 @@ class PokerOpportunist(BehaviorAlgorithm):
                 poker_x = dx / dist
                 poker_y = dy / dist
 
-        # Blend the two behaviors based on weights and energy
-        energy_ratio = fish.energy / fish.max_energy
-
         # Adjust weights based on energy
+        energy_ratio = fish.energy / fish.max_energy
         if energy_ratio < 0.3:
-            # Low energy - prioritize food
-            food_weight = 0.8
-            poker_weight = 0.2
+            food_weight, poker_weight = 0.8, 0.2
         else:
             food_weight = self.parameters["food_weight"]
             poker_weight = self.parameters["poker_weight"]
 
-        # Combine vectors
+        # Blend food and poker attraction
         final_x = food_x * food_weight + poker_x * poker_weight
         final_y = food_y * food_weight + poker_y * poker_weight
         final_len = math.sqrt(final_x * final_x + final_y * final_y)
         if final_len > 0:
             return final_x / final_len, final_y / final_len
 
-        return 0, 0
+        return 0.0, 0.0
 
 
 @dataclass
@@ -536,78 +508,50 @@ class PokerStrategist(BehaviorAlgorithm):
         return cls(rng=rng)
 
     def execute(self, fish: "Fish") -> Tuple[float, float]:
-        import math
-
-        from core.entities import Crab
-
         fish_x, fish_y = fish.pos.x, fish.pos.y
 
-        # Check for predators first
-        nearest_predator = self._find_nearest(fish, Crab)
-        if nearest_predator:
-            dx = nearest_predator.pos.x - fish_x
-            dy = nearest_predator.pos.y - fish_y
-            dist_sq = dx * dx + dy * dy
-            if dist_sq < 14400:  # 120^2
-                dist = math.sqrt(dist_sq)
-                if dist > 0:
-                    return -dx / dist * 1.2, -dy / dist * 1.2
-                return 0, 0
+        flee = _flee_from_predator(self, fish, fish_x, fish_y)
+        if flee is not None:
+            return flee
 
         energy_ratio = fish.energy / fish.max_energy
 
         # Only seek poker if energy is above minimum
         if energy_ratio < self.parameters["min_energy_ratio"]:
-            # Low energy - seek food
-            nearest_food = self._find_nearest_food(fish)
-            if nearest_food:
-                direction = self._safe_normalize(nearest_food.pos - fish.pos)
-                return direction.x * 0.9, direction.y * 0.9
-            return 0, 0
+            return _seek_food_or_idle(self, fish, speed=0.9)
 
-        # OPTIMIZATION: Use spatial query instead of get_agents_of_type
         other_fish = _get_nearby_fish_spatial(fish, 200)
 
         if not other_fish:
-            # No opponents - seek food
-            nearest_food = self._find_nearest_food(fish)
-            if nearest_food:
-                direction = self._safe_normalize(nearest_food.pos - fish.pos)
-                return direction.x * 0.7, direction.y * 0.7
-            return 0, 0
+            return _seek_food_or_idle(self, fish, speed=0.7)
 
         # Strategic target selection based on position and energy
         best_target = None
         best_score = -float("inf")
 
         for target in other_fish:
-            # Calculate distance using raw math
             dx = target.pos.x - fish_x
             dy = target.pos.y - fish_y
             dist_sq = dx * dx + dy * dy
 
-            if dist_sq > 40000:  # 200^2 - Too far
+            if dist_sq > 40000:  # 200^2
                 continue
 
             distance = math.sqrt(dist_sq)
 
-            # Calculate strategic score
+            # Score: prefer optimal distance and energy advantage
             score = 0.0
-
-            # Prefer closer targets (but not too close)
             optimal_distance = 100
             distance_score = 1.0 - abs(distance - optimal_distance) / 150.0
             score += distance_score * self.parameters["position_awareness"]
 
-            # Prefer targets with similar or lower energy
             if hasattr(target, "energy") and target.energy is not None:
                 energy_diff = fish.energy - target.energy
-                if energy_diff > 0:  # We have more energy
+                if energy_diff > 0:
                     score += 0.5 * self.parameters["opponent_tracking"]
-                else:  # They have more energy - risky
+                else:
                     score -= 0.3 * self.parameters["opponent_tracking"]
 
-            # Add some randomness based on aggression variance
             score += self.rng.uniform(
                 -self.parameters["aggression_variance"], self.parameters["aggression_variance"]
             )
@@ -616,19 +560,12 @@ class PokerStrategist(BehaviorAlgorithm):
                 best_score = score
                 best_target = target
 
-        # If we found a good target, pursue it
         if best_target and best_score > 0:
             direction = self._safe_normalize(best_target.pos - fish.pos)
             speed = self.parameters["challenge_speed"]
             return direction.x * speed, direction.y * speed
 
-        # No good targets - balance food and exploration
-        nearest_food = self._find_nearest_food(fish)
-        if nearest_food:
-            direction = self._safe_normalize(nearest_food.pos - fish.pos)
-            return direction.x * 0.8, direction.y * 0.8
-
-        return self.rng.uniform(-0.3, 0.3), self.rng.uniform(-0.3, 0.3)
+        return _seek_food_or_idle(self, fish, speed=0.8)
 
 
 @dataclass
@@ -660,72 +597,46 @@ class PokerBluffer(BehaviorAlgorithm):
         return cls(rng=rng)
 
     def execute(self, fish: "Fish") -> Tuple[float, float]:
-        import math
-
-        from core.entities import Crab
-
         fish_x, fish_y = fish.pos.x, fish.pos.y
 
-        # Check for predators
-        nearest_predator = self._find_nearest(fish, Crab)
-        if nearest_predator:
-            dx = nearest_predator.pos.x - fish_x
-            dy = nearest_predator.pos.y - fish_y
-            dist_sq = dx * dx + dy * dy
-            if dist_sq < 14400:  # 120^2
-                dist = math.sqrt(dist_sq)
-                if dist > 0:
-                    return -dx / dist * 1.2, -dy / dist * 1.2
-                return 0, 0
+        flee = _flee_from_predator(self, fish, fish_x, fish_y)
+        if flee is not None:
+            return flee
 
         # Update behavior mode timer
         self.mode_timer += 1
         if self.mode_timer >= self.mode_duration:
-            # Switch modes randomly
             if self.rng.random() < self.parameters["bluff_frequency"]:
-                modes = ["normal", "aggressive", "passive"]
-                self.current_mode = self.rng.choice(modes)
+                self.current_mode = self.rng.choice(["normal", "aggressive", "passive"])
                 self.mode_duration = self.rng.randint(30, 100)
                 self.mode_timer = 0
 
         # Don't bluff if energy is too low
         if fish.energy < self.parameters["min_energy_to_bluff"]:
-            nearest_food = self._find_nearest_food(fish)
-            if nearest_food:
-                direction = self._safe_normalize(nearest_food.pos - fish.pos)
-                return direction.x, direction.y
-            return 0, 0
+            return _seek_food_or_idle(self, fish)
 
-        # OPTIMIZATION: Use spatial query instead of get_agents_of_type
         other_fish = _get_nearby_fish_spatial(fish, 150)
 
-        # Behavior based on current mode
         if self.current_mode == "aggressive":
-            # Aggressively seek poker
-            nearest_fish, _ = _find_nearest_fish_spatial(fish, 200)
-            if nearest_fish is not None:
-                direction = self._safe_normalize(nearest_fish.pos - fish.pos)
-                speed = 1.0 + self.parameters["aggression_swing"]
-                return direction.x * speed, direction.y * speed
+            result = _seek_fish_or_none(self, fish, 200, 1.0 + self.parameters["aggression_swing"])
+            if result is not None:
+                return result
 
         elif self.current_mode == "passive":
-            # Avoid fish, focus on food
+            # Avoid nearby fish while seeking food
             nearest_food = self._find_nearest_food(fish)
             avoidance_x = 0.0
             avoidance_y = 0.0
 
-            # Calculate avoidance from nearby fish
             for other in other_fish:
                 dx = other.pos.x - fish_x
                 dy = other.pos.y - fish_y
                 dist_sq = dx * dx + dy * dy
-                if dist_sq < 10000 and dist_sq > 0:  # 100^2
+                if 0 < dist_sq < 10000:  # 100^2
                     dist = math.sqrt(dist_sq)
-                    # Direction away from other
                     avoidance_x -= dx / dist
                     avoidance_y -= dy / dist
 
-            # Blend avoidance with food seeking
             avoid_len = math.sqrt(avoidance_x * avoidance_x + avoidance_y * avoidance_y)
             if nearest_food and avoid_len > 0:
                 avoidance_x /= avoid_len
@@ -746,20 +657,12 @@ class PokerBluffer(BehaviorAlgorithm):
                 return direction.x * 0.7, direction.y * 0.7
 
         else:  # normal mode
-            # Balanced approach with unpredictability
             if other_fish and self.rng.random() < 0.6:
-                nearest_fish, _ = _find_nearest_fish_spatial(fish, 150)
-                if nearest_fish is not None:
-                    direction = self._safe_normalize(nearest_fish.pos - fish.pos)
-                    # Add unpredictable speed variation
-                    speed = 0.8 + self.rng.uniform(0, self.parameters["unpredictability"])
-                    return direction.x * speed, direction.y * speed
-
-            # Default to food seeking
-            nearest_food = self._find_nearest_food(fish)
-            if nearest_food:
-                direction = self._safe_normalize(nearest_food.pos - fish.pos)
-                return direction.x * 0.8, direction.y * 0.8
+                speed = 0.8 + self.rng.uniform(0, self.parameters["unpredictability"])
+                result = _seek_fish_or_none(self, fish, 150, speed)
+                if result is not None:
+                    return result
+            return _seek_food_or_idle(self, fish, speed=0.8)
 
         return self.rng.uniform(-0.4, 0.4), self.rng.uniform(-0.4, 0.4)
 
@@ -789,67 +692,42 @@ class PokerConservative(BehaviorAlgorithm):
         return cls(rng=rng)
 
     def execute(self, fish: "Fish") -> Tuple[float, float]:
-        import math
-
-        from core.entities import Crab
-
         fish_x, fish_y = fish.pos.x, fish.pos.y
 
-        # Always flee from predators
-        nearest_predator = self._find_nearest(fish, Crab)
-        if nearest_predator:
-            dx = nearest_predator.pos.x - fish_x
-            dy = nearest_predator.pos.y - fish_y
-            dist_sq = dx * dx + dy * dy
-            if dist_sq < 22500:  # 150^2
-                dist = math.sqrt(dist_sq)
-                if dist > 0:
-                    return -dx / dist * 1.3, -dy / dist * 1.3
-                return 0, 0
+        # Conservative: wider danger zone (150^2) and faster flee (1.3)
+        flee = _flee_from_predator(self, fish, fish_x, fish_y, danger_dist_sq=22500, flee_speed=1.3)
+        if flee is not None:
+            return flee
 
         energy_ratio = fish.energy / fish.max_energy
 
         # Very conservative - only play poker when energy is high
         if energy_ratio < self.parameters["min_energy_ratio"]:
-            # Focus entirely on food
-            nearest_food = self._find_nearest_food(fish)
-            if nearest_food:
-                direction = self._safe_normalize(nearest_food.pos - fish.pos)
-                return direction.x, direction.y
-            return 0, 0
+            return _seek_food_or_idle(self, fish)
 
-        # OPTIMIZATION: Use spatial query with safety distance
-        safety_distance = self.parameters["safety_distance"]
-        other_fish = _get_nearby_fish_spatial(fish, safety_distance)
-
+        # Find favorable matchups (energy advantage)
+        other_fish = _get_nearby_fish_spatial(fish, self.parameters["safety_distance"])
         best_target = None
         best_advantage = 0.0
         energy_advantage_required = self.parameters["energy_advantage_required"]
 
         for other in other_fish:
-            # Check if we have energy advantage
             if hasattr(other, "energy") and other.energy is not None:
                 energy_advantage = fish.energy - other.energy
+                if energy_advantage > energy_advantage_required and energy_advantage > best_advantage:
+                    best_advantage = energy_advantage
+                    best_target = other
 
-                # Only challenge if we have significant energy advantage
-                if energy_advantage > energy_advantage_required:
-                    if energy_advantage > best_advantage:
-                        best_advantage = energy_advantage
-                        best_target = other
-
-        # Only engage if we found a favorable matchup
         if best_target and fish.environment.rng.random() > self.parameters["max_risk_tolerance"]:
             direction = self._safe_normalize(best_target.pos - fish.pos)
             speed = self.parameters["challenge_speed"]
             return direction.x * speed, direction.y * speed
 
-        # Default to safe food seeking
+        # Safe food seeking - avoid close fish while eating
         nearest_food = self._find_nearest_food(fish)
         if nearest_food:
-            # Check if any fish are too close while seeking food (use spatial query)
             close_fish = _get_nearby_fish_spatial(fish, 80)
             if close_fish:
-                # Avoid close fish while seeking food - use raw floats
                 avoidance_x = 0.0
                 avoidance_y = 0.0
                 for f in close_fish:
@@ -882,4 +760,4 @@ class PokerConservative(BehaviorAlgorithm):
                 direction = self._safe_normalize(nearest_food.pos - fish.pos)
                 return direction.x * 0.9, direction.y * 0.9
 
-        return 0, 0
+        return 0.0, 0.0
